Improved GUI and music in Lobby
Added improved textures to the Main menu, aswell as a background. Added an audio source to play music while in the main menu. Also added into the network manager a variable that decides how many players that has to be in the lobby to start the game, mainly for testing purposes